production_trader/requirements.txt
# OANDA v20 API
v20>=3.0.25

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Machine learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# Notifications
python-telegram-bot>=20.0

# Database
sqlite3  # Built-in

# Utilities
requests>=2.31.0
